Update download link for gmp/mpfr/jpeg on osx; older versions are no longer available

+++ .travis/install_prerequisities.sh
@@ -1,24 +1,12 @@
 #!/bin/bash
-sudo apt-get update
-sudo apt-get install -y \
-	curl \
-	git \
-	zsync \
-	xz-utils \
-	libjson-perl \
-	libwww-perl
 
 echo rvm_autoupdate_flag=0 >> ~/.rvmrc
 
 case "$TRAVIS_OS_NAME" in
 linux)
+	sudo apt-get update
 	sudo apt-get install -y \
-		autoconf \
-		automake \
-		build-essential \
-		gcc-4.8 \
-		g++-4.8 \
-		libx11-dev \
+		curl \
 		libosmesa6-dev \
 		libgl1-mesa-dev \
 		libglu1-mesa-dev \
@@ -26,12 +14,26 @@ linux)
 		libsdl-image1.2-dev \
 		libusb-dev \
 		libusb-1.0-0-dev \
-		libudev-dev
-	rm /usr/bin/gcc
-	rm /usr/bin/g++
-	ln -s /usr/bin/gcc-4.8 /usr/bin/gcc
-	ln -s /usr/bin/g++-4.8 /usr/bin/g++
-	gcc -v
+		libudev-dev \
+		zsync \
+		xz-utils \
+		libjson-perl \
+		libwww-perl
+	if test "$emu" = true; then
+		sudo apt-get install -y \
+			git \
+			autoconf \
+			automake \
+			build-essential \
+			gcc-4.8 \
+			g++-4.8 \
+			libx11-dev
+		rm /usr/bin/gcc
+		rm /usr/bin/g++
+		ln -s /usr/bin/gcc-4.8 /usr/bin/gcc
+		ln -s /usr/bin/g++-4.8 /usr/bin/g++
+		gcc -v
+	fi
 	;;
 
 osx)
@@ -74,9 +76,9 @@ osx)
 	# we must install the macports version of the dependencies,
 	# because the brew packages are not universal
 	mkdir src/Unix/MacOSX/external
-	for i in gmp/gmp-6.1.2_0+universal.darwin_16.i386-x86_64.tbz2 \
-		mpfr/mpfr-3.1.5_0+universal.darwin_16.i386-x86_64.tbz2 \
-		jpeg/jpeg-9b_0+universal.darwin_16.i386-x86_64.tbz2; do
+	for i in gmp/gmp-6.1.2_1+universal.darwin_16.i386-x86_64.tbz2 \
+		mpfr/mpfr-4.0.2_0+universal.darwin_12.i386-x86_64.tbz2 \
+		jpeg/jpeg-9c_0+universal.darwin_16.i386-x86_64.tbz2; do
 		f=`basename $i`
 		curl --get https://packages.macports.org/$i --output $f
 		tar -C src/Unix/MacOSX/external --include="./opt/local" --strip-components=3 -xf $f
